Moved SearchBar2024 into common folder, so we can use in WeConnect. Renamed BaseSearchbox to SearchBase.
