+++ src/js/components/Settings/VoterGuideListSearchResults.jsx
@@ -7,7 +7,7 @@ import { SearchTitle } from '../../common/components/Style/FilterStyles';
 import { renderLog } from '../../common/utils/logging';
 import BallotStore from '../../stores/BallotStore';
 import voterGuideSearchPriority from '../../utils/voterGuideSearchPriority';
-import SearchBar2024 from '../Search/SearchBar2024';
+import SearchBar2024 from '../../common/components/Search/SearchBar2024';
 import BallotItemForAddPositions from './BallotItemForAddPositions';
 
 const DelayedLoad = React.lazy(() => import(/* webpackChunkName: 'DelayedLoad' */ '../../common/components/Widgets/DelayedLoad'));
