+++ src/js/config-template.js
@@ -41,12 +41,13 @@ module.exports = {
   // API Keys, some of these are publishable (not secret)
   FACEBOOK_APP_ID: '1097389196952441',
   FULL_STORY_ORG: '',
-  GOOGLE_ADS_TRACKING_ID: '', // Also used for Google TagManager
+  GOOGLE_ADS_TRACKING_ID: '',
   GOOGLE_ANALYTICS_TRACKING_ID: '',
   GOOGLE_MAPS_API_KEY: '',
   GOOGLE_PEOPLE_API_KEY: '',
   GOOGLE_PEOPLE_API_CLIENT_ID: '',
   GOOGLE_RECAPTCHA_KEY: '',
+  GOOGLE_TAG_MANAGER_ID: '',  // for TagManager
   OPEN_REPLAY_PROJECT_KEY: '',
   OPEN_REPLAY_INGEST_POINT: 'https://openreplay.wevote.us/ingest',
   STRIPE_API_KEY: 'pk_test_bWuWGC3jrMIFH3wvRvHR6Z5H',  // This is the publishable key (not secret)
