+++ src/js/pages/Friends/FriendsCurrent.jsx
@@ -6,7 +6,7 @@ import FriendActions from '../../actions/FriendActions';
 import apiCalming from '../../common/utils/apiCalming';
 import { renderLog } from '../../common/utils/logging';
 import FriendList from '../../components/Friends/FriendList';
-import SearchBar2024 from '../../components/Search/SearchBar2024';
+import SearchBar2024 from '../../common/components/Search/SearchBar2024';
 import FriendStore from '../../stores/FriendStore';
 import sortFriendListByMutualFriends from '../../utils/friendFunctions';
 import NoSearchResult from '../../components/Search/NoSearchResult';
